Update smokeTest.php
Extensive changes to customize for my needs. This script will now execute via command line or can be called as a web page for non-CLI users. The #! path was different for my on my Mac (Yosemite), and I needed to add a bunch of stuff to support messages going to the Web via going to the command line. I've added support for straight HREF links and AREAs in an image map that I'm using to simplify/complicate some image touts.

Eventually, I'll be upgrading this with a companion script that will call it against a list of target URLs.